Increase test coverage to model.plausibility.criteria

See merge request psege/rose!111

+++ src/main/java/edu/kit/rose/model/SerializedProject.java
@@ -384,7 +384,7 @@ class SerializedProject {
     @JsonProperty("length")
     private Integer length;
     @JsonProperty("slope")
-    private Integer slope;
+    private Double slope;
     @JsonProperty("laneCount")
     private Integer laneCount;
     @JsonProperty("conurbation")
